- Station UUIDs used by radiobrowser are translated into shorter strings, to work with some older radios - Fixed radiobrowser service stations/byuuid/ REST method name - Version bumped to 1.1.1
